Create DbContext class
Dbcontext class needed to coordinate EF Core functinoality ( CRUD ) for Movie model. the DbContext is derived from Microsoft.EFCore.DbContext.
Creates a DbSet<Movie> property for the entity set. in EF terminology, an entity set typically corresponds to a db table. An entity corresponds to a row in the table.